Code Generator and Validator/requirements.txt
transformers==4.41.2
python-docx==1.1.0
fpdf==1.7.0
streamlit==1.31.0
torch==2.2.1
numpy==1.26.3
sentencepiece==0.1.99
protobuf==4.25.1